refactor(mesh): reorganize file structure
