--- packages/mesh/components/comment/story-comment/index.ts
@@ -1,5 +1,0 @@
-export * from './comment-block'
-export * from './comment-block-item'
-export * from './comment-editor'
-export * from './comment-modal'
-export * from './dropdown-menu'
